refactor: dto 명명 변경
