+++ src/main/java/com/wanted/preonboarding/ticket/domain/dto/request/CreateAlarmPerformanceSeatRequest.java
@@ -1,8 +1,7 @@
-package com.wanted.preonboarding.ticket.domain.dto;
+package com.wanted.preonboarding.ticket.domain.dto.request;
 
 
 import lombok.Builder;
-import lombok.Data;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -10,7 +9,7 @@ import java.util.UUID;
 
 @Getter @Setter
 @Builder
-public class ReservePossibleAlarmCustomerInfoDto {
+public class CreateAlarmPerformanceSeatRequest {
     private UUID performanceId; //공연ID
     private String reservationName; //이름
     private String reservationPhoneNumber; //연락처
